Feat: add refreshtoken logic

+++ Backend/src/main/java/com/solmi/shorket/user/repository/UserRepository.java
@@ -5,10 +5,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
     // Id로 유저 조회
-    Optional<User> findById(Long idx);
+    Optional<User> findById(Integer idx);
 
     // email로 유저 조회
     Optional<User> findByEmail(String email);
